minor: move change stream to own file
